+++ ArduinoMQTTClient/LatchingRelayTwoPole.h
@@ -5,16 +5,16 @@
 
 class LatchingRelayTwoPole : public Output {
 public:
-    LatchingRelayTwoPole(const char* mqttStateTopic, const char* mqttRespondTopic, const unsigned short int setPoleNum, const unsigned short int resetPoleNum);
+    LatchingRelayTwoPole(const char* mqttStateTopic, const char* mqttRespondTopic, const uint8_t setPoleNum, const uint8_t resetPoleNum);
     void action(MQTTDevice::ACTION action, PubSubClient& mqttClient);
 
 private:
-    void togglePole(unsigned short int pinNum);
+    void togglePole(uint8_t pinNum);
     void toggleOn(PubSubClient& mqttClient);
     void toggleOff(PubSubClient& mqttClient);
-    unsigned short int setPoleNum;
-    unsigned short int resetPoleNum;
-    unsigned short int poleHoldTime = 5; // Time to hold a pole for it to register. Datasheet states 3ms.
+    uint8_t setPoleNum;
+    uint8_t resetPoleNum;
+    uint8_t poleHoldTime = 5; // Time to hold a pole for it to register. Datasheet states 3ms.
     bool lastState = LOW;
 };
